% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admissions}
\alias{admissions}
\title{admissions}
\format{
\subsection{\code{admissions}}{

A sample data frame with 400 rows and 4 columns:
\describe{
\item{admit}{Applicant Admitted Flag, 1/0}
\item{gre}{Applicant GRE Score}
\item{gpa}{Applicant GPA}
\item{rank}{Applicant Rank}
}
}
}
\source{
Do not recall.
}
\usage{
admissions
}
\description{
A sample admissions data set
}
\keyword{datasets}
