% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bif_final}
\alias{bif_final}
\title{BIF_Final}
\format{
\subsection{\code{bif_final}}{

A data frame with 2,761 rows and 34 columns:
\describe{
\item{SAMPLEID}{Sample ID}
\item{COHORT}{Cohort}
\item{RA_CODE}{Program or Control Status}
\item{RA_DATE}{Date of Random Assignment}
\item{AGE}{Age at random assignment}
\item{DOB}{Date of Birth}
\item{blmale, blfemale}{Flag for gender}
\item{blmsmarried_lws, blmsmarried_lwos, blmsunmarried_lwp, blmsunmmarried_lwop}{Flags for marital status}
\item{blworkingyes, blworkingno}{Flags for working status}
\item{blcurhrs_1to19, blcurhrs_20to29, blcurhrs_30to34, blcurhrs_35to39, blcurhrs_40to72, blcurhrs_73plus}{Flags for current hours worked}
\item{blhavechild_lt18}{Flag for children under 18}
\item{blhgrade_8ol, blhgrade_9,blhgrade_10,blhgrade_11,blhgrade_12}{Flags for final year of school completed}
\item{bldiplomas_hs, bldiplomas_ged, bldiplomas_tec, bldiplomas_as, bldiplomas_4yr, bldiplomas_md, bldiplomas_non}{Flags for highest level of education}
\item{blotherlang}{Flag for other language spoken at home}
}
}
}
\source{
Sample data created internally
}
\usage{
bif_final
}
\description{
A sample baseline information form data set
}
\keyword{datasets}
