% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterreg_extract.R
\name{clusterreg_extract}
\alias{clusterreg_extract}
\title{Cluster Robust Regression Extract}
\usage{
clusterreg_extract(
  .dataset,
  .dependents,
  .treatment,
  .treatmentNA = NA_character_,
  .cluster_var = NA_character_,
  .covariates = NA_character_,
  .asfactorcovs = NA_character_,
  .subgroup = NA_character_,
  .wt_var = NA_character_,
  .cr_method = "CR2",
  .confintalpha = 0.95,
  .dep_labels = NA_character_,
  .output_path = NA_character_,
  .output_file = NA_character_,
  .inc_trail = TRUE,
  .inc_sample = TRUE
)
}
\arguments{
\item{.dataset}{Input dataframe containing all needed variables}

\item{.dependents}{Character vector of all column names for outcomes.}

\item{.treatment}{Character vector of column name for
.treatment/independent variable}

\item{.treatmentNA}{character string with level of treatment variable to
coerce to NA}

\item{.cluster_var}{Character vector of column name for cluster variable}

\item{.covariates}{Character vector of column names for other
predictors/explanatory variables}

\item{.asfactorcovs}{Character vector of .covariates that should be treated as
factors/categorical variables}

\item{.subgroup}{Character vector of column name to be used to
segment/partition analysis, e.g. sites, cohorts, etc.}

\item{.wt_var}{Character vector of column name to be used as weights
for linear model}

\item{.cr_method}{Character vector of method desired for cluster
adjustment. The default and recommended adjustment is "CR2". Valid values
are: "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3", or "match_SAS".
Please #' see \link[clubSandwich]{vcovCR} for additional details.
The "match_SAS" option is an MDRC specific option to provide the
adjustment as SAS SURVEYREG does, however, it should be noted that this is
not the CR2 adjustment as recommended.}

\item{.confintalpha}{decimal value of desired alpha level for confidence
intervals, i.e. 0.90, 0.95, 0.99. Default is 0.95}

\item{.dep_labels}{Dataframe of labels for dependent variables.
Dataframe must have the columns \code{Dependent} and \code{Labels}}

\item{.output_path}{File path where output excel file should be written.}

\item{.output_file}{Desired file name of output excel file. Should include
.xlsx extension.}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}
}
\value{
Tibble containing results of linear regression analysis on sample
}
\description{
Performs regression, providing cluster-robust standard errors.
}
\details{
Cluster-robust standard-errors are required whenever regression is performed
for studies that have clustered samples. Some examples might be participants
in classrooms, family units, or small-group training. In these situations,
the standard errors from standard regression will be artificially small
because the clusters have less variation within cluster than if they were not
clustered.

The default is estimator in \code{clusterreg_extract()} is CR2. Please see the
function vignette or \link[clubSandwich]{vcovCR} for more information.
}
\examples{
extract <-
clusterreg_extract(.dataset     = sim_data_weighted,
                   .dependents  = c("employed_01", "employed_02"),
                   .treatment   = c("treatment"),
                   .covariates  = c("eduLevel"),
                   .cluster_var = c("sites"),
                   .wt_var      = c("genWeight")
                   )
extract

}
