% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_extract.R
\name{corr_extract}
\alias{corr_extract}
\title{Correlation Extract}
\usage{
corr_extract(
  .dataset,
  .vars = NA_character_,
  .with_vars = NA_character_,
  .by_x = NA_character_,
  .impute = FALSE,
  .method = c("pearson", "spearman", "kendall"),
  .use = c("pairwise.complete.obs", "all.obs", "complete.obs", "everything",
    "na.or.complete"),
  .round = 4,
  .stats = c("mean", "min", "max", "NROW", "sd", "sum"),
  .prob = TRUE,
  .nobs = TRUE,
  .alpha = TRUE,
  .plot = TRUE,
  .output_path = NA_character_,
  .output_file = NA_character_,
  .sheet_name = NA_character_,
  .warnings = FALSE
)
}
\arguments{
\item{.dataset}{Input tibble}

\item{.vars}{Optional, character vector of variables to correlate. Defaults
to all numeric variables not specified in other args (e.g. \code{.with_vars},
\code{.by_x}). Rows in output.}

\item{.with_vars}{Optional, character vector of variables to correlate
with .vars. If none provided, all variables in .vars are correlated with
themselves. Columns in output.}

\item{.by_x}{Optional, string of variable name to separate correlations for}

\item{.impute}{Optional, logical determining if missing values will be
imputed by mean replacement}

\item{.method}{Method to use for correlations, defaults to "pearson"}

\item{.use}{Determines if correlations will be calculated with "pairwise"
deletions or with "complete" data}

\item{.round}{Digits to round to, uses "exact" if NULL}

\item{.stats}{Character vector of other univariate methods to include in
output. Defaults to "mean"}

\item{.prob}{Boolean determining if t-test probability will be displayed. Calculated
using an alpha of 0.05.}

\item{.nobs}{Boolean determining if number of observations will be displayed}

\item{.alpha}{Boolean determining if Cronbach alphas will be calculated}

\item{.plot}{Boolean determining if correlation plot(s) included in Excel}

\item{.output_path}{Output path for excel file}

\item{.output_file}{Output name for excel file}

\item{.sheet_name}{optional, character string with sheet_name for output file}

\item{.warnings}{Boolean determining if Warnings will be printed on run}
}
\value{
Dataframe of correlations and other requested output
}
\description{
Generates correlation extract dataframe using \link[psych]{corr.test}
which is itself a wrapper around \link[stats]{cor} . This function also
optionally attaches the p-value of a two-tailed probability of t and a Cronbach alpha
column if requested.
}
