% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_extract.R
\name{corr_extract_singleton}
\alias{corr_extract_singleton}
\title{Function to generate correlation extract on a particular value of \code{.byx} in
\link[mdrcAnalysis]{corr_extract}}
\usage{
corr_extract_singleton(
  .dataset,
  .vars,
  .with_vars,
  .warnings,
  .method,
  .use,
  .prob,
  .nobs,
  .alpha,
  .stats,
  .round
)
}
\arguments{
\item{.dataset}{Tibble to generate correlations on, only contains numeric data}

\item{.vars}{Optional, character vector of variables to correlate. Defaults
to all numeric variables not specified in other args (e.g. \code{.with_vars},
\code{.by_x}). Rows in output.}

\item{.with_vars}{Optional, character vector of variables to correlate
with .vars. If none provided, all variables in .vars are correlated with
themselves. Columns in output.}

\item{.warnings}{Boolean determining if Warnings will be printed on run}

\item{.method}{Method to use for correlations, defaults to "pearson"}

\item{.use}{Determines if correlations will be calculated with "pairwise"
deletions or with "complete" data}

\item{.prob}{Boolean determining if t-test probability will be displayed. Calculated
using an alpha of 0.05.}

\item{.nobs}{Boolean determining if number of observations will be displayed}

\item{.alpha}{Boolean determining if Cronbach alphas will be calculated}

\item{.stats}{Character vector of other univariate methods to include in
output. Defaults to "mean"}

\item{.round}{Digits to round to, uses "exact" if NULL}
}
\value{
Dataframe containing correlation and, optionally, the univariate statistics
and Cronbach alphas based on passed parameters.
}
\description{
Function to generate correlation extract on a particular value of \code{.byx} in
\link[mdrcAnalysis]{corr_extract}
}
\keyword{internal}
