% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_excel}
\alias{create_excel}
\title{Standard function for repairing and creating file path for excel}
\usage{
create_excel(.output_file = "extract.xlsx", .output_path = NA_character_, .x)
}
\arguments{
\item{.output_file}{Name of exported file; defaults to "extract.xlsx"}

\item{.output_path}{Directory to save Excel; defaults to working directory}

\item{.x}{If object passed, will attempt}
}
\value{
String specifying target file location
}
\description{
Standard function for repairing and creating file path for excel
}
\keyword{internal}
