% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_formula}
\alias{create_formula}
\title{Function for creating formulas}
\usage{
create_formula(
  .TmntVar = NA_character_,
  .CoVars = NA_character_,
  .DepVar = NA_character_
)
}
\arguments{
\item{.TmntVar}{character string with name of desired treatment variable}

\item{.CoVars}{character string with names of desired covariates}

\item{.DepVar}{character string with name of dependent variable.}
}
\value{
tibble with row of sample size data
}
\description{
Returns a formula object from provided variables
}
\details{
Generally used within an inner function which calls only one model at a time.
}
\examples{
\dontrun{
formula <-
create_formula(.TmntVar = .TmntVar,
               .CoVars  = .Covars,
               .DepVar  = .DepVar)}

}
