% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_stars}
\alias{create_stars}
\title{Creates a character vector of p-value stars}
\usage{
create_stars(.p_value)
}
\arguments{
\item{.p_value}{Numeric vector of p-values}
}
\value{
Character vector of stars
}
\description{
Creates a character vector of p-value stars
}
\keyword{internal}
