% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_extract.R
\name{crosstab_clean}
\alias{crosstab_clean}
\title{crosstab_extract helper function to create cross-tab data frame}
\usage{
crosstab_clean(.dataset, .col_var, .row_var, .missing = FALSE)
}
\arguments{
\item{.dataset}{required, data frame with .col_var and .row_var}

\item{.col_var}{required, variable in columns of cross-tabulation}

\item{.row_var}{required, variable in rows of cross-tabulation}

\item{.missing}{optional, Boolean, include missing values, default is FALSE}
}
\value{
tibble with necessary variables and rows
}
\description{
Takes input data frame, subsets for relevant .col_var and .row_var. Checks
for missing values and converts to string, if requested. Outputs data
frame with minimally necessary data for cross-tab.
}
