% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_extract.R
\name{crosstab_extract}
\alias{crosstab_extract}
\title{crosstab_extract main function}
\usage{
crosstab_extract(
  .dataset,
  .col_var,
  .row_vars = NA_character_,
  .subgroup = NA_character_,
  .chisq = FALSE,
  .missing = FALSE,
  .max_lvls = 24,
  .row_labels = NA_character_,
  .inc_trail = TRUE,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.dataset}{required, data frame with .col_var and .row_var}

\item{.col_var}{required, one variable for columns of cross-tabulation}

\item{.row_vars}{optional, character vector of variables in rows of cross-
tabulation. All variables are included if not provided.}

\item{.subgroup}{optional, character vector of variables to group the data by}

\item{.chisq}{optional, Boolean, include chi-square test, default is FALSE}

\item{.missing}{optional, Boolean, include missing values of row variable(s)
in percentage calculations, default is FALSE. Missing counts of column
variable and row variable(s) are always reported in the output table.
Missing values of all variables will be included in chi-square results if
.missing is set to TRUE and chi-square is requested.}

\item{.max_lvls}{optional, maximum number of levels for a variable to be
included in the cross-tabulations. Default is 24.}

\item{.row_labels}{optional, data frame with variable labels. Data frame
should have Variable and Label columns.}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.output_path}{optional, path to output location}

\item{.output_file}{optional, filename for output file}
}
\value{
A tibble of all the cross-tabulations
}
\description{
For creating crosstabs for an entire data frame. Options include creating
crosstabs by subgroup, including missing values, and including chi-square
test results.
}
\examples{
crosstab_extract(.dataset  = bif_final
                ,.col_var  = "RA_CODE"
                ,.row_vars = c("blmale", "bldiplomas_as")
                )

}
