% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_extract.R
\name{crosstab_single}
\alias{crosstab_single}
\title{crosstab_extract helper function: runs crosstab_clean and crosstab_single}
\usage{
crosstab_single(.dataset, .col_var, .row_var, .missing = FALSE, .chisq = FALSE)
}
\arguments{
\item{.dataset}{required, data frame with .col_var and .row_var}

\item{.col_var}{required, variable in columns of cross-tabulation}

\item{.row_var}{required, variable in rows of cross-tabulation}

\item{.missing}{optional, Boolean, include missing values, default is FALSE}

\item{.chisq}{optional, Boolean, request chi-square test, default is FALSE}
}
\value{
clean tibble with all requested statistics
}
\description{
Consolidates both steps for use in main function
}
