% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_extract.R
\name{crosstab_stats}
\alias{crosstab_stats}
\title{crosstab_extract helper which runs the crosstab for a pair of variables}
\usage{
crosstab_stats(.dataset, .col_var, .row_var, .missing = FALSE, .chisq = FALSE)
}
\arguments{
\item{.dataset}{required, data frame with .col_var and .row_var}

\item{.col_var}{required, variable in columns of cross-tabulation}

\item{.row_var}{required, variable in rows of cross-tabulation}

\item{.missing}{optional, Boolean, include missing values, default is FALSE}

\item{.chisq}{optional, Boolean, request chi-square test, default is FALSE}
}
\value{
tibble of single variable cross-tab with chi-square, if requested
}
\description{
Takes input data frame, creates cross-tabulation, calculates desired
frequencies and percentages. Also runs chi-square test, if requested.
}
