% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_extract.R
\name{freq_extract}
\alias{freq_extract}
\title{Generates frequency tables from a given dataset}
\usage{
freq_extract(
  .dataset,
  .freq_vars,
  .max_lvls = 24,
  .subgroup,
  .wt_var,
  .na_rm = FALSE,
  .round = 2,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.dataset}{Required dataframe, Input dataset.}

\item{.freq_vars}{Character vector of variables to process. Defaults to all
variables with fewer than .max_lvls unique values}

\item{.max_lvls}{Optional integer, Used in automatic .freq_vars selection}

\item{.subgroup}{Optional string, Variable name to group frequency by.}

\item{.wt_var}{Optional string, Variable name to weight observations by.}

\item{.na_rm}{Logical, Indicates whether NA values are dropped from
cumulative calculations.}

\item{.round}{Optional integer, Decimal places to round to.}

\item{.output_path}{Optional string, File path of output excel sheet.}

\item{.output_file}{Optional string, File name of output excel sheet.}
}
\value{
Dataframe of frequency table based on input specifications.
}
\description{
Generates frequency tables from a given dataset
}
\examples{
freq_extract(iris, c("Species", "Sepal.Length"))
freq_extract(iris)
freq_extract(mtcars, .subgroup = "cyl", .max_lvls = 5, .na_rm = TRUE)
}
\author{
Audrey Yu
}
