% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetreg_extract.R
\name{hetreg_extract}
\alias{hetreg_extract}
\title{Heteroskedastic Robust Regression Extract}
\usage{
hetreg_extract(
  .dataset,
  .dependents,
  .treatment,
  .treatmentNA = NA_character_,
  .covariates = NA_character_,
  .asfactorcovs = NA_character_,
  .subgroup = NA_character_,
  .wt_var = NA_character_,
  .hcc_method = "HC3",
  .confintalpha = 0.95,
  .dep_labels = NA_character_,
  .output_path = NA_character_,
  .output_file = NA_character_,
  .inc_trail = TRUE,
  .inc_sample = TRUE
)
}
\arguments{
\item{.dataset}{Input dataframe containing all needed variables}

\item{.dependents}{Character vector of all column names for outcomes.}

\item{.treatment}{Character vector of column name for
.treatment/independent variable}

\item{.treatmentNA}{character string with level of treatment variable to
coerce to NA}

\item{.covariates}{Character vector of column names for other
predictors/explanatory variables}

\item{.asfactorcovs}{Character vector of .covariates that should be treated as
factors/categorical variables}

\item{.subgroup}{Character vector of column name to be used to
segment/partition analysis, e.g. sites, cohorts, etc.}

\item{.wt_var}{Character vector of column name to be used as weights
for linear model}

\item{.hcc_method}{Character vector of method desired for
heteroskedasticity adjustment. Valid values are: "HC3", "const", "HC", "HC0",
"HC1", "HC2", "HC4", "HC4m", and "HC5". Please #' see \link[sandwich]{vcovHC}
for additional details.}

\item{.confintalpha}{decimal value of desired alpha level for confidence
intervals, i.e. 0.90, 0.95, 0.99. Default is 0.95}

\item{.dep_labels}{Dataframe of labels for dependent variables.
Dataframe must have the columns \code{Dependent} and \code{Labels}}

\item{.output_path}{File path where output excel file should be written.}

\item{.output_file}{Desired file name of output excel file. Should include
.xlsx extension.}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}
}
\value{
Tibble containing results of linear regression analysis on sample
}
\description{
Performs regression, providing heteroskedasticity-robust standard errors.
}
\details{
Heteroskedasticity, or heterogeneity of variance, is when the variance of
observations is not uniform across the distribution. That is, observations in
different areas of the distribution have different variance. If this occurs
in the data, then the estimation of the standard errors using standard
regression will be incorrect. To avoid incorrect standard errors, various
estimators can be used.

At MDRC, it is recommended to use heteroskedasticity-consistent standard-
error estimators whenever regression is performed as the adjusted standard
errors \emph{generally} have benign effects if the sample size is large enough.
The default is estimator in \code{hetreg_extract()} is HC3. Please see the
function vignette or \link[sandwich]{vcovHC} for more information.
}
\examples{
extract <-
hetreg_extract(.dataset    = sim_data_weighted,
               .dependents = c("employed_01", "employed_02"),
               .treatment  = c("treatment"),
               .covariates = c("eduLevel"),
               .wt_var     = c("genWeight")
               )
extract

}
