% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_extract.R
\name{ht_extract}
\alias{ht_extract}
\title{Homogeneity Test Extract}
\usage{
ht_extract(
  .dataset,
  .subgroup,
  .dependents,
  .treatment,
  .covariates,
  .asfactorcovs = NA_character_,
  .confintalpha = 0.95,
  .inc_trail = TRUE,
  .inc_sample = TRUE,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.dataset}{Input data}

\item{.subgroup}{String specifying column encoding subgroup rows belong to}

\item{.dependents}{Character vector of dependent variables to run models on}

\item{.treatment}{String specifying treatment variable}

\item{.covariates}{Character Vector of covariates to run model on}

\item{.asfactorcovs}{Character vector of covariates to be run as
factors/categorical variables}

\item{.confintalpha}{decimal value of desired alpha level for
confidence intervals. Default is 0.95}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}

\item{.output_path}{file path to output location}

\item{.output_file}{output file name}
}
\value{
Tibble displaying model outputs over all combinations of
dependents and subgroup values
}
\description{
The HT statistic is a test of homogeneity of effect size.
The HT statistic is used to check whether impacts for subgroups are
statistically different.
}
