% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_extract.R
\name{lm_extract}
\alias{lm_extract}
\title{Generate dataframe of linear model extracts across a set of dependent variables}
\usage{
lm_extract(
  .dataset,
  .dependents,
  .treatment,
  .covariates = NA_character_,
  .asfactorcovs = NA_character_,
  .confintalpha = 0.95,
  .inc_trail = TRUE,
  .inc_sample = TRUE,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.dataset}{data frame name providing all necessary variables}

\item{.dependents}{character vector of all column names for outcomes}

\item{.treatment}{character vector of column name(s) for .treatment/independent variable}

\item{.covariates}{character vector of column names for other predictors/explanatory variables}

\item{.asfactorcovs}{character vector of .covariates that should be treated as factors/categorical variables}

\item{.confintalpha}{decimal value of desired alpha level for confidence intervals, i.e. 0.90, 0.95, 0.99. Default is 0.95}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}

\item{.output_path}{optional, path to output location}

\item{.output_file}{optional, filename for output file}
}
\value{
data frame with one row per dependent with desired regression statistics
}
\description{
Generate dataframe of linear model extracts across a set of dependent variables
}
\examples{
# Functional Sample Call
# Copy outside of function and run after function

# Convert COHORT to a factor for analysis
FinalBIF <- dplyr::mutate(bif_final, CohortFactor = factor(COHORT))

# Create vector of dependent variables
## Using `setdiff()` to create vector of all variables except those listed
FinalBIF_dependents <- names(FinalBIF[setdiff(names(FinalBIF),
                                              c("SAMPLEID",
                                                "RA_CODE",
                                                "CohortFactor",
                                                "COHORT",
                                                "RA_DATE",
                                                "AGE",
                                                "DOB",
                                                "blcurhrs_73plus"))])
FinalBIF_.treatment  <- "RA_CODE"
FinalBIF_.covariates <- "CohortFactor"

extract <- lm_extract(FinalBIF,
                      FinalBIF_dependents,
                      FinalBIF_.treatment,
                      FinalBIF_.covariates,
                      .confintalpha = .9)

extract
}
\author{
Melvin Gutierrez
}
