% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{match_arg_null}
\alias{match_arg_null}
\title{Match an argument to a character vector with an option for NULL return}
\usage{
match_arg_null(.arg, .choices, .several_ok = FALSE)
}
\arguments{
\item{.arg}{A character vector (of length one unless \code{.several_ok} is TRUE) or
NULL which will return NULL}

\item{.choices}{A character vector of candidate values, often missing. See
\link[base:match.arg]{base::match.arg} for details.}

\item{.several_ok}{logical specifying if \code{.arg} should be allowed to have more
than one element.}
}
\value{
The unabbreviated version of the exact or unique partial match if exists;
otherwise an error is signalled if \code{.several_ok} is \code{FALSE} as per default.
When \code{.several_ok} is \code{TRUE}, then all elements of \code{.arg} must match
and these are returned. If \code{NULL} is passed, then \code{NULL} is returned
}
\description{
This is equivalent to \link[base:match.arg]{base::match.arg} with a few key
differences:
\itemize{
\item Explicitly passing \code{NULL} will return NULL instead of the first match
\item If \code{.several_ok} is TRUE, then ALL elements of \code{.arg} must have
a partial match
\item If \code{.several_ok} is TRUE, then matched elements can only be used once.
}
This is done to facilitate easy argument verification while still allowing for an
intuitive way specify that none of the options should be taken
}
