% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrcAnalysis_package.R
\docType{package}
\name{mdrcAnalysis-package}
\alias{mdrcAnalysis}
\alias{mdrcAnalysis-package}
\title{mdrcAnalysis: Functions for Analysis at MDRC}
\description{
A package for analysis functions commonly used at MDRC. Current functions include \link[mdrcAnalysis]{freq_extract}, \link[mdrcAnalysis]{mean_extract}, \link[mdrcAnalysis]{corr_extract}, \link[mdrcAnalysis]{crosstab_extract}, \link[mdrcAnalysis]{ttest_extract}, \link[mdrcAnalysis]{lm_extract}, \link[mdrcAnalysis]{hetreg_extract}, \link[mdrcAnalysis]{clusterreg_extract}, \link[mdrcAnalysis]{ht_extract}, and \link[mdrcAnalysis]{mixed_extract}.
}
\author{
\strong{Maintainer}: Melvin Gutierrez \email{Melvin.Gutierrez@mdrc.org}

Authors:
\itemize{
  \item Benjamin Bui \email{Benjamin.Bui@mdrc.org}
  \item Julie Glew \email{Julie.Glew@mdrc.org}
  \item Zarni Htet \email{Zarni.Htet@mdrc.org}
  \item Dorota Rizik \email{Dorota.Rizik@mdrc.org}
}

}
\keyword{internal}
