% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_extract.R
\name{mean_extract}
\alias{mean_extract}
\title{mean extract function}
\usage{
mean_extract(
  .dataset,
  .subgroup = NA_character_,
  .quantile = FALSE,
  .inc_sample = TRUE,
  .inc_trail = TRUE
)
}
\arguments{
\item{.dataset}{required, the data set with necessary variables}

\item{.subgroup}{optional, character vector of variables to group the data by}

\item{.quantile}{optional, request quantiles, default is FALSE}

\item{.inc_sample}{RTU parameter for sample size row inclusion, default is TRUE}

\item{.inc_trail}{RTU parameter for trail inclusion, default is TRUE}
}
\value{
tibble with summary statistics for all variables
}
\description{
mean extract function
}
\examples{
mean_extract(
            .dataset = sim_data_robust_reg,
            .subgroup = "treatment",
            .quantile = TRUE
            )


}
