% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_check.R
\name{missing_check}
\alias{missing_check}
\title{Searches for missing proxies across a dataset}
\usage{
missing_check(.data, .factordb = NULL, .missing_proxy = c(77, 88, 99))
}
\arguments{
\item{.data}{Dataframe}

\item{.factordb}{Optional, factor dataframe listing all factors}

\item{.missing_proxy}{Vector of missing proxy values}
}
\description{
Missing data is a frequent occurrence; however, missing data
frequently appears in forms more complex than just an NA. Often descriptions
of the cause of the missing data are present which are represented in the
data with a string or number. Once some of these proxies are identified,
they can be searched through across a dataset to see other potential missings
}
