% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extract.R
\name{mixed_by_compare.default}
\alias{mixed_by_compare.default}
\title{Helper Function for extracts that use emmmeans::emmeans()}
\usage{
\method{mixed_by_compare}{default}(
  .dataset,
  .model,
  .emmeans,
  .treatment,
  .reference_level,
  .dependent
)
}
\arguments{
\item{.model}{lm or lmerMod class object}

\item{.emmeans}{emmGrid class object}

\item{.reference_level}{String specifying reference value to use in .subgroup}

\item{.dependent}{String specifying dependent variable}

\item{.subgroup}{String specifying by_group used in emmeans}
}
\description{
Helper Function for extracts that use emmmeans::emmeans()
}
\keyword{internal}
