% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extract.R
\name{mixed_extract}
\alias{mixed_extract}
\title{Generate mixed extract given a dataframe object.}
\usage{
mixed_extract(
  .dataset,
  .dependents,
  .rhs = NA_character_,
  .treatment = NA_character_,
  .reference_level = NA_character_,
  .fixed_effects = NA_character_,
  .random_intercept = NA_character_,
  .alternate = c("cluster", "aggregate", "skip"),
  .output_path,
  .output_file
)
}
\arguments{
\item{.dataset}{Data to create model on. Must inherit class data.frame}

\item{.dependents}{Vector of dependent variables to create extracts on}

\item{.rhs}{Optional string to use as right hand side lme4::lmer formula.
Required for more complex models}

\item{.treatment}{character vector for .treatment/independent variable}

\item{.reference_level}{Optional string specifying which value of .treatment is
used as reference value. Defaults to first value.}

\item{.fixed_effects}{NOT IN USE Vector of fixed effects if .rhs not specified}

\item{.random_intercept}{NOT IN USE String specifying variable to use for random
intercept If vector of length > 1 passed, assumed to be mlm of > 2 levels.
Grouping structure of levels taken from order of vector with first character
being lowest level.}

\item{.alternate}{NOT IN USE Alternate method if mixed model fails to converge}

\item{.output_path}{File path to write excel to}

\item{.output_file}{File name to write excel to}
}
\value{
dataframe displaying relevant mixed model statistics
}
\description{
Generate mixed extract given a dataframe object.
}
