% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extract.R
\name{mixed_extract_aggregate}
\alias{mixed_extract_aggregate}
\title{Alternative mlm extract generated using aggregated lm instead of mlm}
\usage{
mixed_extract_aggregate(
  .dataset,
  .dependent,
  .rhs,
  .subgroup,
  .reference_level
)
}
\arguments{
\item{.dataset}{data frame being passed by user}

\item{.dependent}{character value of desired outcome variable}

\item{.rhs}{model specification for predictors}

\item{.subgroup}{Required, String specifying .subgroup}

\item{.reference_level}{Required, String specifying reference value of .subgroup}
}
\value{
Tibble row summarizing aggregated linear model results
}
\description{
Extract used as a multi-level model substitute for
categorical dependent variables. Does not follow work-flow process of other
functions because this process is so unique
}
