% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extract.R
\name{mixed_extract_cluster}
\alias{mixed_extract_cluster}
\title{Alterantive mlm extract generated using clustered lm instead of mlm}
\usage{
mixed_extract_cluster(.dataset, .dependent, .rhs, .subgroup, .reference_level)
}
\arguments{
\item{.dataset}{data frame being passed by user}

\item{.dependent}{character value of desired outcome variable}

\item{.rhs}{model specification for predictors}

\item{.subgroup}{Required, String specifying .subgroup}

\item{.reference_level}{Required, String specifying reference value of .subgroup}
}
\value{
Tibble row summarizing clustered linear model results
}
\description{
Alterantive mlm extract generated using clustered lm instead of mlm
}
