% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_extract.R
\name{mixed_extract_mixed.data.frame}
\alias{mixed_extract_mixed.data.frame}
\title{Specific method for mixed_extract, default mixed}
\usage{
\method{mixed_extract_mixed}{data.frame}(
  .dataset,
  .dependent,
  .rhs,
  .treatment = NA_character_,
  .reference_level = NA_character_
)
}
\arguments{
\item{.dataset}{Input data}

\item{.dependent}{String specifying dependent variable}

\item{.rhs}{Right hand side of mixed model formula}

\item{.treatment}{String specifying .treatment}

\item{.reference_level}{String specifying reference value of \code{.treatment}}
}
\value{
Tibble row summarizing mixed model results
}
\description{
Specific method for mixed_extract, default mixed
}
