% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_na}
\alias{paste_na}
\title{Paste with NA suppression}
\usage{
paste_na(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{One or more objects that can be converted to character vectors.}

\item{sep}{Character string to separate terms}

\item{collapse}{Optional string to separate terms}
}
\value{
Result of base::paste call ignoring all NULL, NA, or 0-length args
If all arguments of ... are excluded, then return NA_character_
}
\description{
Paste with NA suppression
}
