% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustreg_extract.R
\name{robustreg_extract}
\alias{robustreg_extract}
\title{Robust Regression Extract}
\usage{
robustreg_extract(
  .dataset,
  .TmntVar,
  .DepVars,
  .TreatmentNA = NA_character_,
  .DepLabels = NA_character_,
  .CoVars = NA_character_,
  .asfactorcovs = NA_character_,
  .RobustSE = TRUE,
  .ClusterVar = NA_character_,
  .StrataVar = NA_character_,
  .ByX = NA_character_,
  .WtVar = NA_character_,
  .Total = NA_character_,
  .Rate = NA_character_,
  .Intercept = TRUE,
  .OtherModelOptions = NA_character_,
  .RptProc = NA_character_,
  .RptOut = NA_character_,
  .output_path = NA_character_,
  .output_file = NA_character_,
  .Replace = NA_character_,
  .Round = NA_character_,
  .RoundProb = NA_character_,
  .inc_trail = TRUE,
  .inc_sample = TRUE
)
}
\arguments{
\item{.dataset}{Input dataframe containing all needed variables}

\item{.TmntVar}{Character vector of treatment variables}

\item{.DepVars}{Character vector of dependent variables}

\item{.TreatmentNA}{Value of treatment variable to coerce to NA}

\item{.DepLabels}{Labels for dependent variables, not enabled}

\item{.CoVars}{Character vector of covariates}

\item{.asfactorcovs}{character vector of .covariates that should be treated as
factors/categorical variables}

\item{.RobustSE}{Boolean determining if should be done with robust SE}

\item{.ClusterVar}{Character vector of variables to cluster by if
clustered SE desired}

\item{.StrataVar}{Not yet enabled}

\item{.ByX}{Variable with values for analysis running by}

\item{.WtVar}{Weights for linear model}

\item{.Total}{Not yet enabled}

\item{.Rate}{Not yet enabled}

\item{.Intercept}{Boolean determining if the implied intercept term should be included}

\item{.OtherModelOptions}{Not yet enabled}

\item{.RptProc}{Not yet enabled}

\item{.RptOut}{Not yet enabled}

\item{.output_path}{Excel file path}

\item{.output_file}{Excel file name}

\item{.Replace}{Not yet enabled}

\item{.Round}{Not yet enabled}

\item{.RoundProb}{Not yet enabled}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}
}
\value{
Tibble containing results of linear regression analysis on sample
}
\description{
Performs robust regression analysis.
}
