% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sample_size}
\alias{sample_size}
\title{Sample Size Function}
\usage{
sample_size(.dataset, .treatment = NA_character_)
}
\arguments{
\item{.dataset}{entire data set}

\item{.treatment}{specify treatment group}
}
\value{
tibble with row of sample size data
}
\description{
Returns the sample size of the entire data set and by treatment group variables
}
\examples{

mdrcAnalysis:::sample_size(
           .dataset = sim_data_robust_reg,
           .treatment = "treatment"
            )

}
