% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_three_levels_missing}
\alias{sim_three_levels_missing}
\title{sim_three_levels_missing}
\format{
\subsection{\code{sim_data}}{

A data frame with 323 rows and 21 columns:
\describe{
\item{cities}{cities ID}
\item{population}{population of city}
\item{sites}{Site ID}
\item{individuals}{ID}
\item{age_raw, age_scaled}{Age}
\item{eduLevel}{Flag for education level}
\item{pre_income_raw, pre_income_scaled, pre_income_raw_25pct}{income}
\item{tanf_receipt, snap_receipt}{Benefits}
\item{gender}{gender}
\item{numChild}{children under 18}
\item{maritalStatus}{Marital Status}
\item{treatment}{Program or Control Status}
\item{probs_01, probs_02}{probabilities}
\item{employed_01, employed_02}{employment status}
\item{favAnimals}{favorite Animals}
}
}
}
\source{
Sample data created internally
}
\usage{
sim_three_levels_missing
}
\description{
A simulated data set with missing data
}
\keyword{datasets}
