% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_repair_latex}
\alias{str_repair_latex}
\title{Creating latex safe text}
\usage{
str_repair_latex(.x)
}
\arguments{
\item{.x}{Character vector to coerce}
}
\value{
Character vector with special characters formatted to be LaTeX safe
}
\description{
Creating latex safe text
}
