% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trail}
\alias{trail}
\title{Call Trail}
\usage{
trail(call = NA_character_)
}
\arguments{
\item{call}{parameter to provide results of \link[mdrcAnalysis]{match.call.defaults}}
}
\value{
Tibble containing:
filepath (if script run in RStudio)
username
datetime with timezone
(Note: timezone will be EST because that's where MDRC Servers Run)
R Version
packages and version numbers
@keywords internal
}
\description{
Generate dataframe of session information
This information can be joined to extracts for documentation purposes
}
