% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_extract.R
\name{ttest_extract}
\alias{ttest_extract}
\title{T-Test Extract}
\usage{
ttest_extract(
  .df,
  .classvar,
  .depvars,
  .depvar_labels,
  .byx = NULL,
  .alternative = c("two.sided", "less", "greater"),
  .mu = 0,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.df}{Required, Input dataframe}

\item{.classvar}{Required, Categorical independent variable. Must have
exactly 2 different values}

\item{.depvars}{Optional, names of numeric dependent variables. Defaults to
all numeric variables}

\item{.depvar_labels}{Optional, character vector of labels to use for
.depvars}

\item{.byx}{Optional, variable to to split and group results by}

\item{.alternative}{Alternative hypothesis, defaults to two sided test}

\item{.mu}{Value indicating "true" difference in mean}

\item{.output_path}{Optional, path to output file location}

\item{.output_file}{Optional, output file name}
}
\value{
Tibble describing difference within dependent variables between
binary categorical variable
}
\description{
Outputs frequencies, percents, and results of a t-test for two independent
samples to an excel sheet and dataframe
DOES NOT compare a mean to a constant value, perform a paired comparison,
or extract confidence levels
}
