% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_extract.R
\name{ht_extract}
\alias{ht_extract}
\title{Homogeneity of Observed Effects Extract}
\usage{
ht_extract(
  .dataset,
  .output_path = NA_character_,
  .output_file = NA_character_
)
}
\arguments{
\item{.dataset}{Input data, an extract with subgroups from
\link[mdrcAnalysis]{lm_extract}, \link[mdrcAnalysis]{hetreg_extract},
\link[mdrcAnalysis]{clusterreg_extract}}

\item{.output_path}{file path to output location}

\item{.output_file}{output file name}
}
\value{
Tibble displaying model outputs over all combinations of
dependents and subgroup values
}
\description{
The HT statistic is a test of homogeneity of observed effects (impacts).
The HT statistic is used to check whether impacts across subgroups are
statistically different.
}
\details{
Greenberg, D., Meyer, R. H., & Wiseman, M. (1994). Multisite Employment
and Training Program Evaluations: A Tale of Three Studies. Industrial and
Labor Relations Review, 47(4), 679–691. https://doi.org/10.2307/2524666

The paper is available in the QMG SharePoint site. Please reach out if there
are questions.
}
