% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_extract.R
\name{lm_extract}
\alias{lm_extract}
\title{General linear regression extract}
\usage{
lm_extract(
  .dataset,
  .dependents,
  .treatment,
  .covariates = NA_character_,
  .asfactorcovs = NA_character_,
  .subgroup = NA_character_,
  .wt_var = NA_character_,
  .confintalpha = 0.95,
  .dep_labels = NA_character_,
  .output_path = NA_character_,
  .output_file = NA_character_,
  .inc_trail = TRUE,
  .inc_sample = TRUE
)
}
\arguments{
\item{.dataset}{Input dataframe containing all needed variables}

\item{.dependents}{Character vector of all column names for outcomes.}

\item{.treatment}{Character vector of column name for
.treatment/independent variable}

\item{.covariates}{Character vector of column names for other
predictors/explanatory variables}

\item{.asfactorcovs}{Character vector of .covariates that should be treated as
factors/categorical variables}

\item{.subgroup}{Character vector of column name to be used to
segment/partition analysis, e.g. sites, cohorts, etc.}

\item{.wt_var}{Character vector of column name to be used as weights
for linear model}

\item{.confintalpha}{decimal value of desired alpha level for confidence
intervals, i.e. 0.90, 0.95, 0.99. Default is 0.95}

\item{.dep_labels}{Dataframe of labels for dependent variables.
Dataframe must have the columns \code{Dependent} and \code{Labels}}

\item{.output_path}{optional, path to output location}

\item{.output_file}{optional, filename for output file}

\item{.inc_trail}{RTU parameter to include trail in output}

\item{.inc_sample}{RTU parameter to include sample size row in output}
}
\value{
data frame with one row per dependent with desired regression statistics
}
\description{
Performs regression across a set of dependent variables.
}
\details{
NOTE: This function does NOT provide heteroskedasticity-robust standard
errors. In general, staff will want to use heteroskedasticity-robust standard
errors, as in the \link[mdrcAnalysis]{hetreg_extract} function.
}
\examples{
# Functional Sample Call

extract <-
lm_extract(.dataset    = sim_data_weighted,
           .dependents = c("employed_01", "employed_02"),
           .treatment  = c("treatment"),
           .covariates = c("eduLevel"),
           .wt_var     = c("genWeight")
           )

extract

}
\author{
Melvin Gutierrez
}
